Trabalho 3 funcionando, aparentemente

+++ viewport.h
@@ -21,6 +21,6 @@ public:
 	}
 
 private:
-	int xMin = 0.0, yMin = 0.0, xMax = 300.0, yMax = 300.0;
+	float xMin = 0.0, yMin = 0.0, xMax = 300.0, yMax = 300.0;
 
 };
